api/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
azure-storage-blob
shapely
requests
werkzeug
sendgrid>=6.9.7